+++ test/Audit.IntegrationTest/Audit.IntegrationTest.csproj
@@ -22,7 +22,6 @@
     <ProjectReference Include="..\..\src\Audit.NET.Kafka\Audit.NET.Kafka.csproj" />
     <ProjectReference Include="..\..\src\Audit.NET\Audit.NET.csproj" />
     <ProjectReference Include="..\..\src\Audit.NET.Udp\Audit.NET.Udp.csproj" />
-    <ProjectReference Include="..\..\src\Audit.NET.MongoDB\Audit.NET.MongoDB.csproj" />
     <ProjectReference Include="..\..\src\Audit.NET.SqlServer\Audit.NET.SqlServer.csproj" />
     <ProjectReference Include="..\..\src\Audit.NET.AzureStorage\Audit.NET.AzureStorage.csproj" />
     <ProjectReference Include="..\..\src\Audit.NET.MySql\Audit.NET.MySql.csproj" />
